+++ src/app/(dashboard)/layout.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from '@/components/SideBar'
+import { Sidebar } from '@/components/dashboard/SideBar'
 
 interface DashboardLayoutProps {
   children: React.ReactNode
